reorganize some parts
other part
