+++ fabric-registry-sync-v0/src/main/java/net/fabricmc/fabric/api/event/registry/RegistryAttribute.java
@@ -48,6 +48,7 @@ public enum RegistryAttribute {
 
 	/**
 	 * For implementation purposes only!
+	 * Returned key values are subject to change.
 	 */
 	@Deprecated
 	@ApiStatus.Internal
